feat(APM/docker): docker image and docker compose

+++ src/lib/apm/AgentStore.ts
@@ -9,14 +9,16 @@ import { AGENT } from './Agent.js';
 class AgentStore {
 	axios: AxiosInstance;
 	baseURL: string = ServerConfig.apm.agentStore.baseURL;
-	apiKey: string;
+	username: string;
+	password: string;
+	sessionToken: string;
 
-	constructor(baseURL?, apiKey?) {
+	constructor(baseURL?, sessionToken?) {
 		this.axios = axios.create();
 
 		this.setBaseURL(baseURL);
-		this.setApiKey(apiKey);
-		if (!this.apiKey) {
+		this.setSessionToken(sessionToken);
+		if (!this.sessionToken) {
 			this.readCachedAuthFile();
 		}
 	}
@@ -35,7 +37,12 @@ class AgentStore {
 			url: '/agentstore/agent/login',
 			data: { username, password },
 		});
-		this.setApiKey(response.data.sessionToken);
+		if (response.data.error) {
+			throw new EmpError(response.data.error, response.data.message);
+		}
+		this.username = username;
+		this.password = password;
+		this.setSessionToken(response.data.sessionToken);
 		await this.saveToCachedAuthFile();
 		return response.data;
 	}
@@ -80,10 +87,14 @@ class AgentStore {
 
 		// save to tmp
 		const tmp_dir = await AGENT.getTMPWorkDirCreate();
+		console.log("save to tmp",tmp_dir)
 		const md5 = await AGENT.saveUploadFile(tmp_dir, payload.file);
 		// read from tmp
+		console.log("read from tmp",md5)
 		const filename = `${md5}.tar.gz`;
+
 		const filepath = path.join(tmp_dir, filename);
+		console.log("filepath",filepath)
 
 		const response = await this.axios({
 			method: 'POST',
@@ -104,13 +115,13 @@ class AgentStore {
 		return response.data;
 	}
 
-	setApiKey(apiKey: string) {
-		if (apiKey) {
-			this.apiKey = apiKey;
+	setSessionToken(sessionToken: string) {
+		if (sessionToken) {
+			this.sessionToken = sessionToken;
 		}
 
-		if (this.apiKey) {
-			this.axios.defaults.headers.common['Authorization'] = `Bearer ${this.apiKey}`;
+		if (this.sessionToken) {
+			this.axios.defaults.headers.common['Authorization'] = `Bearer ${this.sessionToken}`;
 		}
 	}
 	setBaseURL(baseURL: string) {
@@ -150,7 +161,7 @@ class AgentStore {
 
 		const { auth } = await fs.readJson(filepath);
 
-		this.setApiKey(auth?.agentstore?.apiKey);
+		this.setSessionToken(auth?.agentstore?.sessionToken);
 	}
 	async saveToCachedAuthFile() {
 		const filepath = path.resolve(ServerConfig.apm.localRepositoryDir, 'apm.json');
@@ -169,7 +180,11 @@ class AgentStore {
 			fileJSON = await fs.readJson(filepath);
 		}
 
-		Object.assign(fileJSON.auth.agentstore, { apiKey: this.apiKey });
+		Object.assign(fileJSON.auth.agentstore, {
+			username: this.username,
+			password: this.password,
+			sessionToken: this.sessionToken,
+		});
 
 		await fs.writeJson(filepath, fileJSON, { spaces: 4 });
 	}
@@ -202,6 +217,10 @@ class AgentStore {
 		return await this.moveToAuthorAgentDir(untarDir, agentStoreAgent);
 	}
 	async untar(filepath, outputDir) {
+		if(await fs.exists(outputDir)){
+			return;
+		}
+		
 		await fs.ensureDir(outputDir);
 		console.log('untar', filepath, '=>', outputDir);
 		await child_process.exec(`tar zxvf ${filepath}`, {
@@ -221,6 +240,9 @@ class AgentStore {
 		);
 		const agentVersionDir = path.resolve(agentNameDir, agentStoreAgent.version);
 		await fs.ensureDir(path.dirname(agentNameDir));
+		if(await fs.exists(agentVersionDir)){
+			return agentVersionDir;
+		}
 		await fs.move(folder, agentVersionDir);
 
 		return agentVersionDir;
